feature/redux -  code refactor - add service - fetch data from fake server
